requirements.txt
streamlit==1.37.1
python-docx==1.1.2
pymupdf==1.24.9
python-dotenv==1.0.1
requests==2.32.3
